Move profile types into /ui/src/types. Update links
